// web/src/api/models/DtoSubscriptionFreeTrialModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SubTracker API
 * This api provide HTTPRest endpoints for the application SubTracker.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@mistribe.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * @Description Number of free trial days remaining (null if no trial or trial expired)
 * @export
 * @interface DtoSubscriptionFreeTrialModel
 */
export interface DtoSubscriptionFreeTrialModel {
    /**
     * 
     * @type {Date}
     * @memberof DtoSubscriptionFreeTrialModel
     */
    endDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof DtoSubscriptionFreeTrialModel
     */
    startDate: Date;
}

/**
 * Check if a given object implements the DtoSubscriptionFreeTrialModel interface.
 */
export function instanceOfDtoSubscriptionFreeTrialModel(value: object): value is DtoSubscriptionFreeTrialModel {
    if (!('endDate' in value) || value['endDate'] === undefined) return false;
    if (!('startDate' in value) || value['startDate'] === undefined) return false;
    return true;
}

export function DtoSubscriptionFreeTrialModelFromJSON(json: any): DtoSubscriptionFreeTrialModel {
    return DtoSubscriptionFreeTrialModelFromJSONTyped(json, false);
}

export function DtoSubscriptionFreeTrialModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): DtoSubscriptionFreeTrialModel {
    if (json == null) {
        return json;
    }
    return {
        
        'endDate': (new Date(json['end_date'])),
        'startDate': (new Date(json['start_date'])),
    };
}

export function DtoSubscriptionFreeTrialModelToJSON(json: any): DtoSubscriptionFreeTrialModel {
    return DtoSubscriptionFreeTrialModelToJSONTyped(json, false);
}

export function DtoSubscriptionFreeTrialModelToJSONTyped(value?: DtoSubscriptionFreeTrialModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'end_date': value['endDate'].toISOString(),
        'start_date': value['startDate'].toISOString(),
    };
}