// web/src/api/models/DtoAmountModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SubTracker API
 * This api provide HTTPRest endpoints for the application SubTracker.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@mistribe.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * @Description Custom price for this subscription
 * @export
 * @interface DtoAmountModel
 */
export interface DtoAmountModel {
    /**
     * 
     * @type {string}
     * @memberof DtoAmountModel
     */
    currency: string;
    /**
     * 
     * @type {DtoAmountModel}
     * @memberof DtoAmountModel
     */
    source?: DtoAmountModel;
    /**
     * 
     * @type {number}
     * @memberof DtoAmountModel
     */
    value: number;
}

/**
 * Check if a given object implements the DtoAmountModel interface.
 */
export function instanceOfDtoAmountModel(value: object): value is DtoAmountModel {
    if (!('currency' in value) || value['currency'] === undefined) return false;
    if (!('value' in value) || value['value'] === undefined) return false;
    return true;
}

export function DtoAmountModelFromJSON(json: any): DtoAmountModel {
    return DtoAmountModelFromJSONTyped(json, false);
}

export function DtoAmountModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): DtoAmountModel {
    if (json == null) {
        return json;
    }
    return {
        
        'currency': json['currency'],
        'source': json['source'] == null ? undefined : DtoAmountModelFromJSON(json['source']),
        'value': json['value'],
    };
}

export function DtoAmountModelToJSON(json: any): DtoAmountModel {
    return DtoAmountModelToJSONTyped(json, false);
}

export function DtoAmountModelToJSONTyped(value?: DtoAmountModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'currency': value['currency'],
        'source': DtoAmountModelToJSON(value['source']),
        'value': value['value'],
    };
}